Fix ITM logging in the examples.
This code needs a modified stm32g4x.cfg that creates stm32g4x.tpiu. The
file stm32f4x.cfg from OpenOCD can be used as an example, as it already
has the appropriate line.

The code used to steal all peripherals in the ITM logging implementation,
which caused the examples to panic when calling take(). Now, the code just
creates a new ITM instance.