Support for I2C (#26)
* Add support for I2C.

Most of this code is taken from stm32g0xx-hal, with only minor changes to
the way pins are initialized and to the interaction with RCC.

* Improve the I2C documentation.

* Fix compiler error for devices without I2C4.

* Add an I2C/MPU6050 example (the timing is still broken).

* Fix the I2C bit rate.

The std::max() meant that scll was always set to 255, resulting in bit
rates that were far too slow. Also, the documentation states that the
clock is divided by PRESC+1, not by 2^(PRESC+1).