FlashWriter.write requires additional user input
- FlashWriter.write used to automatically pad incoming data such that is
was divisible by 8, a requirement of the STM32 flash hardware. A
colleague suggested this could cause confusion during subsequent
writes. FlashWriter.write now forces the user to acknowledge this
with the `force_data_padding` flag. If true, the data will be padded
automatically to be divisible by 8. If false and the data is not
divisible by 8, an error is thrown.
- Added new error ArrayMustBeDivisibleBy8
- Modified the flash example based on the above updates